backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
drf-spectacular==0.29.0
ultralytics==8.3.229
pillow==12.0.0
google-genai==1.51.0
google-auth==2.43.0